Migration on server startup #862hxqa2z

+++ BaCa2/course/models.py
@@ -72,7 +72,7 @@ class Submit(models.Model):
     submit_date = models.DateTimeField(auto_now_add=True)
     source_code = models.FileField(upload_to=SUBMITS_DIR)
     task = models.ForeignKey(Task, on_delete=models.CASCADE)
-    usr = models.IntegerField()  # TODO: user id
+    usr = models.FloatField()  # TODO: user id
     final_score = models.FloatField(default=-1)
 
     def __str__(self):
